fix atexit issue and ray hostcall reorg
