+++ wasm/src/engine/wasmtime_engine/mod.rs
@@ -18,6 +18,9 @@ use anyhow::Result;
 use std::collections::HashMap;
 use wasmtime::{Engine, FuncType, Instance, Linker, Module, Store, Val, ValRaw, ValType};
 
+mod data;
+use data::*;
+
 use super::wasmedge_engine;
 
 pub struct WasmtimeEngine {
@@ -239,61 +242,3 @@ impl WasmtimeInstance {
 }
 
 impl WasmInstance for WasmtimeInstance {}
-
-pub fn to_wasmtime_raw_value(val: &WasmValue) -> wasmtime::ValRaw {
-    match val {
-        WasmValue::I32(v) => wasmtime::ValRaw::i32(*v),
-        WasmValue::I64(v) => wasmtime::ValRaw::i64(*v),
-        WasmValue::F32(v) => wasmtime::ValRaw::f32(*v),
-        WasmValue::F64(v) => wasmtime::ValRaw::f64(*v),
-        WasmValue::V128(v) => wasmtime::ValRaw::v128(*v),
-        WasmValue::ExternRef(v) => wasmtime::ValRaw::externref(*v),
-        WasmValue::FuncRef(v) => wasmtime::ValRaw::funcref(*v),
-    }
-}
-
-pub fn to_wasmtime_value(val: &WasmValue) -> Val {
-    match val {
-        WasmValue::I32(v) => Val::I32(*v),
-        WasmValue::I64(v) => Val::I64(*v),
-        WasmValue::F32(v) => Val::F32(*v),
-        WasmValue::F64(v) => Val::F64(*v),
-        WasmValue::V128(v) => Val::V128(*v),
-        _ => unimplemented!(),
-    }
-}
-
-pub fn from_wasmtime_raw_value(ty: &WasmType, val: &wasmtime::ValRaw) -> WasmValue {
-    match ty {
-        WasmType::I32 => WasmValue::I32(val.get_i32()),
-        WasmType::I64 => WasmValue::I64(val.get_i64()),
-        WasmType::F32 => WasmValue::F32(val.get_u32()),
-        WasmType::F64 => WasmValue::F64(val.get_u64()),
-        WasmType::V128 => WasmValue::V128(val.get_v128()),
-        WasmType::ExternRef => WasmValue::ExternRef(val.get_externref()),
-        WasmType::FuncRef => WasmValue::FuncRef(val.get_funcref()),
-    }
-}
-
-pub fn from_wasmtime_value(ty: &WasmType, val: &Val) -> WasmValue {
-    match ty {
-        WasmType::I32 => WasmValue::I32(val.unwrap_i32()),
-        WasmType::I64 => WasmValue::I64(val.unwrap_i64()),
-        WasmType::F32 => WasmValue::F32(val.unwrap_f32().to_bits()),
-        WasmType::F64 => WasmValue::F64(val.unwrap_f64().to_bits()),
-        WasmType::V128 => WasmValue::V128(val.unwrap_v128()),
-        _ => unimplemented!(),
-    }
-}
-
-pub fn wasmtime_type(ty: &WasmType) -> wasmtime::ValType {
-    match ty {
-        WasmType::I32 => ValType::I32,
-        WasmType::I64 => ValType::I64,
-        WasmType::F32 => ValType::F32,
-        WasmType::F64 => ValType::F64,
-        WasmType::V128 => ValType::V128,
-        WasmType::ExternRef => ValType::ExternRef,
-        WasmType::FuncRef => ValType::FuncRef,
-    }
-}
